WIP: Split raster functions to their own file. Improve test coverage.
