+++ ldn/classify.py
@@ -33,10 +33,9 @@ from typing_extensions import Annotated
 from ldn.aws_credentials import get_write_client, get_write_session, make_write_function
 from ldn.geomad import AwsStacTask as Task
 from ldn.grids import get_gridspec
-from ldn.raster import PrefixedS3ItemPath
+from ldn.raster import GEOMAD_BANDS, PrefixedS3ItemPath, calculate_indices, load_dem_terrain, scale_offset_landsat
 from ldn.utils import (
     AWS_REGION,
-    GEOMAD_BANDS,
     GEOMAD_VERSION,
     PREDICTION_DATASET_ID,
     PREDICTION_VERSION,
@@ -45,11 +44,8 @@ from ldn.utils import (
     SOURCE_COOP_PUBLIC_URL,
     WGS84,
     LdnError,
-    calculate_indices,
     get_analysis_epsg,
     get_geomad_stac_geoparquet_url,
-    load_dem_terrain,
-    scale_offset_landsat,
 )
 
 logger = logging.getLogger(__name__)
